+++ packages/account/src/providers/transaction-summary/index.ts
@@ -6,4 +6,4 @@ export * from './operations';
 export * from './get-transaction-summary';
 export * from './assemble-transaction-summary';
 export * from './receipt';
-export * from './calculate-transaction-fee';
+export * from './calculate-tx-fee-for-summary';
